compare Finalized

+++ unesco_oer/templates/content/6b_tpl.php
@@ -295,8 +295,14 @@ if ($check == FALSE) {
     } else {
         $existingContent = $content->getContentByContentID($moduleid);
     }
-    echo '<div class="heading2"><h2 class="greyText">' . $existingContent->getTitle() . ' </div><br>';
-    echo $existingContent->showReadOnlyInput();
+   if ($existingContent != FALSE){
+                     
+                     echo '<div class="heading2"><h2 class="greyText">' .$existingContent->getTitle() . ' </div><br>';      
+                     echo  $existingContent->showReadOnlyInput();
+          
+                 }
+                 else 
+                     echo '<div class="heading2"><h2 class="greyText">Module/Curriculum Not Available </div><br>'; 
 }
 ?>
